vpu5: handle the end of stream
This patch includes 4 changes. The first change is to detect an EOS
flag in the input buffer. The second puts an EOS flag into the final
output buffer. The third outputs the 2 NALs for indicating the end
of stream, that are the End of Sequence and the End of Stream.
The final adds code to finalize the VPU M/W.