vpu5: Make sure the component always flushes input buffers (Experimental)
If a NAL unit crosses a buffer boundary and a flush buffer request is
recieved, the component will be in a state when it is both requesting
a new buffer and processing an existing buffer.  In this case the
buffer being processed will not be flushed when it should be, causing
an error. Instead we flush anything in the queue of buffers currently
being processed.