vpu5: Remove semaphore around interrupt handler
The implementation assumed that interrupts were only processed during
mcvdec_decode_picture, but this is not the case.  The decoder functions
mcvdec_end_decoder and mcvdec_flush_buff also wait for interrupts to
signal the end of vlc operation.  Since the semaphore is not needed in
the single instance case its use has been removed (variable definitions
still remain).

**This patch removes support for multiple decoder instances. (Support
will be reenabled when its working properly)