vpu5: avcenc: follow the de-facto port parameter setting mannar
The OMXIL spec. never defines which and how input/output port
parameters should be used, however, Some major OMXIL video encoder
clients has a mannar in that width, height, framerate for raw image
should be set in the input port parameter, and bitrate for encoded
stream should be set in the output port parameter.
This patch changes process of reflecting the port setting into the
middle-ware parameters for following the manner.