vpu5: avcenc: fix type of xFramerate to follow the OMXIL spec.
The bellagio ffmpeg component and bellagio examples assumes that
the xFramerate parameter shoule be specified with an interger,
such 25 for 25.0 fps, however, the OMXIL spec. defines that
xFramerate is Q16 fixed point type.
This patch treats xFramerate as a Q16 format value rather than
integer.
MEMO: omxvideenctest should be fixed for the change.