vpu5: Set MERAM to use higher address range on IPMMUI init
The default setting for the MERAM has the ICBs and the IPMMUI using
the same address range starting at 0x80000000.  This patch sets a
mode bit in the MERAM to use the range starting at 0xC0000000 instead.
Moving the IPMMUI range would be preferable, but the IPMMUI range is fixed.