vpu5: avcdec: Move settings for maxVideoParameters and maxVideoInstances to constructor
The settings for maxVideoParameters and maxVideoInstances were set in
SetInternalVideoParameters.  This function is called both on init by the
constructor and when the component role is set.  The Android framework
gives us a nice opportunity to set parameters through the overloaded
createInstance method, but the component role is set after this via a call to
SetParameter.  If we re-init these settings when the component role changes
we lose the opportunity to set them when the component is created.