vpu5ha: Select between UseBuffer and AllocateBuffer implementations
The USE_BUFFER_MODE flag can be set or unset to change the supported
allocation mode.  This is necessary because the AllocateBuffer and
UseBuffer implementations use different methods to return data to
the application.  The AllocateBuffer version can write directly to
the output buffers, while the UseBuffer version needs to use a DMA
or copying.