vpu5ha: Allow for adjusting of output buffer stride alignment
Some downstream devices may have specific requirements for the width of the
output buffers.  While this doesn't strictly adhere to the OMXIL requirements,
we allow for aligning the output buffer width (in pixels) to be a multiple of
a certain value.

Set the alignment by defining OUTPUT_BUFFER_ALIGN.  Leave it undefined for normal
operation.