vpu5ha: Properly handle filler data NAL unit parsing
The end of picture was not being properly marked when
filler data NAL units were present in the data stream

Also, since filler data NAL units don't have any picture data,
we have to make sure not to assign their timestamps
(which are usually 0) to the decoded picture. Instead we
use the timestamp from the buffer containing the NAL
that holds the actual picture data.