vpu5: Always unlock vpu after mcvenc_encode_picture
It should not be necessary to hold the vpu lock until the
callback from the middleware is received, even if not in
B-picture mode.

Furthermore if the component is shutdown while the lock is held,
it will never be realeased and the component will not be able to
process data if/when it is re-initialized.

With this change the component can never be shutdown while the vpu
lock is held.