avcdec: Limit AVC level in software readable mode
Software readable mode is used primarily for creating thumbnails, which
may occur in multiple threads at the same time.  We usually don't need
fast decode for this and we don't have enough memory to run multiple
decodes in parallel at the higher bitrate modes.  Cap the AVC level at 4.0
when using this mode.