vpu5: mpeg4 parse: Allow for any sized input picture to be processed
The parser uses fixed sized buffers to pass input data to the VPU. Since
the entire fixed sized picture must fit into one of these buffers, if the
input picture was too large to fit in the buffer, it could not be decoded.
Now, if the input buffer is too large, the VPU buffer is reallocated to
a larger size and the previously truncated data is added.
This only occurs for the largest pictures in high bit-rate streams.
The buffer resizing requires an additional copy operation to be performed,
which may result in reduced performance on streams with bitrates over about 60Mbps.