vpu5: Set default maximum resolutions based on available VPU memory
On systems with limited VPU memory, setting the default maximum
decode resolution and profile level too high with consume most
of the available memory before decoding starts.  This means that even
lower resolution streams cannot be decoded.
Set the default resolution and profile level  to a lower value for
systems with VPU memory <= 64MiB. The default can be overridden with
a call to SetParameter if needed.