vpu5: Update available buffer check before VPU output
In some casees, the VPU output may undergo post-filtering before
being output to the OMX client buffers.  In this case we want to
make sure that the post-filtered output, not the direct VPU output
is syncronized so that it is not overwritten before the client
has a chance to read it.

With post-filtering enabled, trying to syncronize the VPU output to
the client buffers results in the VPU locking one set of buffers, while
the client unlocks an entirely different set of buffers, resulting in
a deadlock.