vpu: Use clip region to properly set up decoded image size
The nFrameWidth value should always correspond to the size
of the video being decoded.  The nStride value will correstpond
to the buffer stride.

Likewise the clip region should be available to an OMX client
that wishes to use that in addition to the nFrameWidth and
nStride to determine the video dimensions.