vpu5: Adjust the PMB size to 512MB
This will allow the PMB T/L inversion to work as long as the VPU memory
region does not cross a 512MB boundary.  For systems with < 512MB, any
memory configuration will work. For systems with > 512MB, care must be
taken if the total CMA (carveout) memory size is > (total mem - 512MB)