vpu5: Reduce memory usage when in thumbnail mode
When making thumbnails only 1 frame is processed at a time.  As such
we shouldn't need a huge VBV buffer to buffer the stream data.
Reduce the VBV buffer size to be a multiple of the MSS (maximum stream
size) when we process thumbnails.  This allows for processing multiple
thumbnails in parallel without allocating hundreds of megabytes of memory.