vpu5: avc parse: Try to access the first MB flag only when it should be present
The first macroblock flag as defined in the slice header may not
be present in all Nal units. Specifally End of stream and End of sequence
Nal units do not define RBSP bits, and typically appear as the
last Nal unit of the stream.  Trying to access a non-existent byte to
find the first_mb flag can lead to a segmentation fault in this instance.