vpu5: Allow for reallocation of output buffers on Android
If a portSettingsChanged callback is made to Android from the component,
Android will try to reallocate the output buffers based on the
new buffer size.  This will happen while the component is still in the
Executing state by disabling the output buffer and reissuing the
UseBuffer commands.  Update the status check make sure that the DMAC setup
is reentrant.