vpu5: Reduced the number of calling the uio_malloc
Inspected the mediaserver by oProfile to find out a performance bottleneck at
playing a video. The oProfile report showed that the uio_malloc used a lot of
cpu resource. After debugging, identified that the copyNalData in
the libshvpu5avc.so called the uio_malloc many times while playing a video.
To avoid it, cached a memory which is allocated via the uio_malloc in
the libshvpu5avc when the memory was freed. Therefore, the cached memory is
reused when the uio_malloc is called again.
As a result, the memory cache reduced the number of calling the uio_malloc, and
a CPU utilization via the mediaserver was decreased by about 5 %.